requirements2.txt
autopep8==1.3.5
beautifulsoup4==4.6.0
boto==2.48.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.0
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.6
django-background-task==0.1.8
django-background-tasks==1.1.13
django-compat==1.0.15
django-storages==1.6.6
gunicorn==19.8.1
idna==2.7
kiwisolver==1.0.1
lxml==4.2.1
matplotlib==2.2.2
networkx==2.1
numpy==1.14.5
pandas==0.23.1
Pillow==5.1.0
psycopg2==2.7.5
pycodestyle==2.4.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
requests==2.19.1
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
toolz==0.9.0
urllib3==1.23
whitenoise==3.3.1
